pyproject.toml
[project]
name = "opensearch-genai-sdk"
version = "0.1.0"
description = "OpenSearch AI Observability SDK — OTEL-native tracing, scoring, and evaluation for LLM applications"
authors = [
    { name = "OpenSearch Contributors" },
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "opentelemetry-api>=1.20.0,<2",
    "opentelemetry-sdk>=1.20.0,<2",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0,<2",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0,<2",
]

[project.optional-dependencies]
aws = [
    "botocore>=1.29.0",
    "requests-aws4auth>=1.2.0",
]
evals = [
    "autoevals>=0.0.50",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
all = [
    "opensearch-genai-sdk[aws,evals]",
]

[project.urls]
Repository = "https://github.com/vamsimanohar/opensearch-genai-sdk"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opensearch_genai_sdk"]

[tool.hatch.build.targets.sdist]
exclude = ["examples/", "tests/", "docker-compose.yaml", "otel-collector-config.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
